refactor: colocation
close #34
